Set OpenGL to byte mode for unpacking
At the same time, this disables GL_UNPACK_CLIENT_STORAGE_APPLE. These
two settings are orthogonal, but Gargoyle does not adhere to the
requirements of this Apple extension:

“Clients using this extension are agreeing to preserve a texture's image
data for the life of the texture.  The life of the texture is defined,
in this case, as the time from first issuing the TexImage3D, TexImage2D
or TexImage1D command, for the specific texture object with the
UNPACK_CLIENT_STORAGE_APPLE pixel storage parameter set to TRUE, until
the DeleteTextures command or another TexImage command for that same
object.”

From https://registry.khronos.org/OpenGL/extensions/APPLE/APPLE_client_storage.txt.

In Gargoyle's case, the texture data comes from gli_image_rgb; the data
is a pointer to the internal buffer there. When gli_image_rgb is
resized, the internal buffer will be modified, very possibly in a way
that makes previous pointers to it invalid. That will happen *before*
texture deletion, violating the requirements of the extension. Since
nothing really happens between the resize and redrawing of the texture,
this has probably not been the source of any issues, but it's still
invalid.

Thanks to Petter Sjölund for this fix.