Throw a "proper" exception type
This really doesn't matter, since the exception is just being used to
signal some error; the actual error itself is not important. But it's a
little weird to be throwing some type that comes from libjpeg, so just
use std::exception instead.

Maybe in the future this can throw the actual error message and log it.