Improve lookup of interpreters
This makes a couple changes.

First, if an interpreter for a specific game is provided in the config
file, unconditionally use that interpreter now. Previously Gargoyle
would ignore that setting if the file was (or appeard to be) a Blorb.

Second, don't check extensions for Blorb files. In the past, Gargoyle
exclusively used extensions to determine which interpreter to run. A
while back, file probing was added, with extensions used as a fallback.
But the extension for Blorb was checked before probe information was
used, which resulted in the Blorb launcher being called for .blb files,
even if they were probed as another file format, causing failure.
There's no need to check the Blorb extension, because Blorb files can be
probed. If it didn't look like a Blorb file to the probing code, it will
be futile to try to load it as a Blorb, even with a .blorb extension.

The steps now are something like this:

1. Is an interpreter supplied in the config file? If so, run that
   interpreter.
2. Does it have a Blorb header? If so, run it as a Blorb file, failing
   if it can't properly be loaded.
3. Was it probed as a specific file format? If so, run the associated
   interpreter.
4. If the file extension is known, run the associated interpreter.
   Otherwise, fail.