Remove discrete TADS and Z-machine versions
The same interpreter is used for both TADS2 and TADS3; and the same
intepreter is used for all Z-machine versions.

I don't know why there was ever a TADS split. For the Z-machine, Nitfol
used to be used for V6 games, and Frotz the others, but now there's just
a single Z-machine interpreter. Nothing is gained by separating out
these versions.