Require the regular font variation to exist
Without this, it's possible to get a mix of the Gargoyle fallback fonts
and the user fonts (e.g. if only italic exists, the user's font will be
used for italic and Gargoyle for the rest).

If somebody really wants a mix of fonts, they can use the propr, propb,
etc settings.