Impose Gargoyle's style on Gargoyle code in CheapGlk
This is an odd combination: upstream cheapglk is used, mostly, but there
are some Gargoyle additions to it. These additions were more-or-less in
the same coding style as CheapGlk, except for the fact that C++ is used
in some places, where CheapGlk is C (in fact, the whole CheapGlk has
been "converted" to C++, but with almost no modifications).

Since C++ is being used, it's a bit strange to follow C conventions, so
now the Gargoyle additions are the same style as the rest of the
Gargoyle code. The CheapGlk parts can mostly be ignored, since they're
not supposed to be touched. Any changes will come from upstream. In
short, the CheapGlk code is treated mostly as a black box, so failing to
match its style isn't really important. That's the story I'm going with,
anyway.