Don't remove channels during shutdown
This is not an ideal solution (nor are any of the other uses of
gli_exit): if any interpreter calls exit() directly (rather than
glk_exit()), then this whole thing is sidestepped. But it's a tricky
problem, dealing with order of destruction of static objects. Keep
hacking around for now, and hopefully a better solution will eventually
be discovered.