Put sysqt's View and Window in the garglk namespace
This is yet another concession to the byzantine nature of C++.

It is undefined behavior in C++ to have two different types with the
same name. Not just in the same file, but across the whole program! And
C++ doesn't concern itself with what, exactly, a program is. That leads
to a really ugly problem here: even though libgarglk is a "separate"
entity, it's not really: if there is a type in libgarglk with the same
name as a type in one of the interpreters, that's undefined.

For basic types this probably isn't an issue, but it's _definitely_ an
issue for classes. On modern Unix compilers, at least, if you compile a
file containing a class, using no optimization, the class methods will
be included in the object file, and will be global. Since the name of a
class determines the name of its constructor, that means that two
totally separate classes, but with the same name, will generate
identical external constructor names. Which one will be used? Who knows!

Once optimizations are enabled, the names are no longer external. So
while this is still undefined behavior as far as C++ is concerned, at
least there's a fairly stable way around it.

The change here is reactive, not proactive. It comes from a real
problem: Bocfel has a Window class, and Gargoyle (when using Qt) has a
Window class. If optimization is turned off, then there will be _two_
Window::Window() constructors, and the first one the linker finds will
be used. This led to breakage: in testing, Bocfel's Window objects were
being initialized with Gargoyle's Window constructor, with unsurprising
results (i.e. crashing).

The full solution is to put everything in a namespace, but that's going
to take a bit of effort, since the codebase was originally C and thus
had no concept of namespaces. For now, do it in one are that is _known_
to be a problem.