Don't install a man page when using DIST_INSTALL
In fact, don't use GNUInstallDirs at all in this configuration, and
don't set $INTERPRETER_INSTALL_DIR.

Mac sets UNIX, but if Mac's doing DIST_INSTALL it is _not_ doing a
normal "make install". But if(UNIX) would still pass, and try to install
a man page to /usr/local.

If DIST_INSTALL is set, then the normal installation procedure is
upended, so none of this is relevant. This logic should probably move
into garglk/CMakeLists.txt, but this is a quick fix for now.