Update Plus interpreter
Testing the unreleased X-Men fragment against the original interpreter
uncovered a surprising number of problems with my implementation.

Two opcodes, 126 and 127, unused by the earlier games, were not working correctly.

One condition,5, would fail right away instead of continue checking against other
conditions it was ORed with.

The original interpreter keeps an extra copy of the input tokens around, which is
copied to the corresponding counters. The earlier games didn't need this, but X-Men
does, so I changed the implementation to match the original.

Also, some of the images in the Apple 2 version of Questprobe 3 were mis-labeled.