Provide warnings for invalid config file entries
Prior to this Gargoyle would either ignore or pull into range invalid
values in the config file, and the only real way to know you got it
wrong was if your config change didn't do anything. Now it dumps errors
to stderr, which isn't particularly helpful for anybody using a GUI to
launch, but it's a start. The framework can be used to provide better
error messages in the future, e.g. in a dialog box or similar.

At the same time, all "float" values have been converted to "double".
Some config values were float, others double, so moving them all to
double just simplifies things.