Add "save scrollback" feature
With control/command-shift-S, the current scrollback buffer can be
written to a file. Because Glk supports an arbitrary number of windows,
there is no "definite" scrollback to save. Gargoyle tries the focused
window first; if that's not a textbuffer, it then searches through all
texstbuffers and chooses the one with the largest scrollback buffer.
Hopefully this covers the vast majority of cases.