Allow picture scaling to be selected for Magnetic Scrolls
This is a bit hacky due to the way Gargoyle works. The Gargoyle config
file doesn't directly affect games at all, so the choice is either to
allow the setting in the Gargoyle config or to create a separate config
for Magnetic Scrolls.

Because the upstream Magnetic Scrolls doesn't already have a config
file, it's less invasive to just add a new command-line argument that
allows scaling, since Gargoyle _does_ allow arguments to be passed to
interpreters, although that support had to be expanded a bit since it
originally was very limited, meant only to support one thing for AGT.