Support images in SCARE
This works by writing a temporary PIC0 file, displaying it, then
removing it. A future Gargoyle extension might allow memory-based image
display, rendering this unnecessary, but it works well enough for the
time being. The PIC0 file must be in the game directory, though, so if
that's not writable, this won't work.

Many games display an image on startup then clear the screen, which
obliterates the image. The official Adrift runner, by default, has a
separate window-manager window showing the image, so it doesn't matter
if the screen is cleared. However, you can configure it to display
inline images, in which case, just like in SCARE, they disappear,
although unlike Gargoyle, you can scroll back past cleared screens in
the Adrift runner.

The Adrift runner also supports a Glk-like window on the top or right
holding the graphic (plus a map), so theoretically SCARE could be
modified to support both inline and windowed images as well, but for the
time being, just do inline. This loses title screen images, but Gargoyle
never had that before, so it's not a change, and games with actual
inline images (such as "The PK Girl") now work.