Support saving window position on Mac
This isn't 100% perfect: the height will be a bit off when saved, so it
doesn't come back in exactly the same spot. This is probably fixable but
it won't be trivial, since it has to do with the window being updated
after a signal is sent, so a mechanism for saying "store window position
after first update signal" needs to be implemented. And, hopefully, the
Mac interface will be retired soon enough in favor of Qt which does all
the work for us already.