Handle newlines properly when pasting
Prior to this, the Qt and Mac interfaces behaved different to each other
when pasting in line input mode, and both in a manner that most people
would consider wrong.

The Mac code would strip out newlines, pasting multiple lines all at
once.

The Qt code would send newlines appropriately as a Glk "return" keycode,
but after that keycode, the window was no longer in "accept input" mode,
which meant that all incoming keys were discarded. All the rest of the
keys arrive too early, before the next line input is started, and are
discarded as a result.

Char input is similar, but there _any_ characters is sufficient to stop
the event and cause the rest of the paste to disappear.

Now, if a paste occurs when there's no input event, the pasted
characters are buffered and injected when there _is_ an input event.