Use the new garglk_add_resource_from_file function for Hugo/Scare
This is still not ideal: the official Adrift runner can display images
in a separate window manager window or in a split window, and a lot of
games expect this, clearing the main window after displaying an image.
Since Gargoyle displays images inline, they'll never be seen. But at
least some games' images will work.

I'll probably add an option at some point to split the window for
images, but for the time being, they are all inline. Another possibility
is to intercept "clear window" calls and just print several newlines,
but that's a bit of an overstep into a game's intent, maybe.