Patch for Transporter
This game tries to read a property from an object without checking if
it is valid in a routine at 0x4bd1. This will be called with a
negative value when trying to ASK the wizard about things, and with the
number typed when trying to TYPE a number, which may be larger than the
number of objects.

We fix this by checking if the argument is a valid object.