Don't build FrankenDrift in Fedora
This _does_ work on my local system, so it's some configuration issue
with the runner, apparently; for now, just don't do FrankenDrift in the
RPM build. These generated RPMs are not really meant to be used, but are
rather a smoke test for the RPM build, so it's not critical that
Frankendrift be included here. Ideally it would, and will be resurrected
if it can be gotten to work in the future.