Remove unnecessary #undef
WIN32 is a legacy macro defined by some Windows headers, and shouldn't
really be used today. _WIN32 is set by the compiler, and is the proper
macro to check if you're actually on Windows.

But Hugo _appears_ to expect you to define WIN32 manually, i.e. it's
unrelated to the Windows WIN32 macro. If that's the case, then if GARGLK
is defined, then WIN32 won't be, as Gargoyle never defines WIN32.

On the other hand, if it's intending to check for the obsolete Windows
WIN32 macro, that doesn't matter either: since Windows headers aren't
being included, it won't be defined.

In any case, this ought to be unnecessary, and the comment isn't really
germane, either: Hugo doesn't try to use WinGlk if WIN32 is set.