Handle the case of no provided story
Gargoye used to just assume that argv[argc - 1] is the story file; and
things "worked", sort of, because if argc was 1, then it'd be checking
argv[0] as the story file, but because bracketed sections in the config
file accept both a story and interpreter, it was just redundant with the
explicit argv[0] check.

Similarly gamepath is now optional, so if there is no argv[argc - 1], it
won't even be consulted. This has an effect: gamepath used to be "" when
no story was provided, and the code would just blindly assume it was a
filename: for example, replacing the extension with ".ini" would result
in searching for a file called simply ".ini".