Revert type changes to upstream Magnetic
The Magnetic code has loads of pointer type mismatches, essentially
assuming that type8s (signed char) is compatible with char, which it
isn't. char can be unsigned (and is, on aarch64), and even where it's
signed, char and signed char are distinct types, and you cannot convert
between pointers to the two without casting.

Originally I used "char" for "type8s" since it is generally used
synonymously with "char" in the code, but there are indeed some areas
where it _needs_ to be signed, breaking things on aarch64. While
compilers are allowed to reject invalid pointer conversions between
differently-signed integers of the same rank, in reality they don't. GCC
doesn't even warn about such conversions by default, although the
standard mandates a diagnostic.

Clang does warn, so for Clang (and GCC, for good measure), use #pragmas
to disable this particular warning. Eventually if a strict compiler
comes along we'll have to just litter the code with casts, but this is
good enough for now.