Install fallback fonts to data directory alongside Unifont
Gargoyle needs to be able to find several different fonts:

• Fallback monospace and proportional, in case the user's selected fonts
  can't be found.
• Unifont for missing glyphs.

Until now, the fallback fonts (on Unix) were placed in /usr/share/fonts
by default, i.e. the system font directory. But that's not really ideal:
even though Gargoyle asks for the fonts by name and they are
custom-named fonts, so won't clash with the system, there's no
requirement for them to actually _be_ system fonts. Nobody is going to
use Gargoyle's fonts outside of Gargoyle. Even if they wanted to,
Gargoyle's fonts are just Charis SIL and Go Mono, so those fonts could
be used directly.

All that _needs_ to happen is for Gargoyle to be able to find these
fonts when necessary. It used to embed the fonts directly in libgarglk,
but due to licensing issues (the OFL isn't compatible with the GPL) they
had to be installed separately, and _somehow_ discoverable.

Unifont was always placed in a "data" directory, e.g.
/usr/share/io.github.garglk/Gargoyle on Unix. It's not installed as a
system font (the user very possibly already has it, anyway), and just
needs to be discoverable by Gargoyle. So it's a "data file" as far as
Gargoyle is concerned.

Given that, there's no reason for there to be two separate font
locations, so the fallback fonts are now also installed to the data
directory. Theoretically there's no reason for them to be called
"Gargoyle Serif" and "Gargoyle Mono" anymore, but since previous
Gargoyle versions shipped with those names, it's probably too late to do
anything. Perhaps it's better this way, in case the default fonts need
to change again: the "Gargoyle" names stick around regardless of the
underlying font name.

In short, fonts are simplified by placing all needed fonts in the same
location now.