Be more specific when looking up the data directory
There are two things the winappdata() method was used for:

1. Looking up theme paths
2. Finding Unifont

But these are distinct operations. When looking for themes, the system
themes should be used (as a base), but the user should also be able to
add themes. As such, the concept of XDG's data directories, being
configurable, makes perfect sense: allow the user to specify where he
wants his data directories to be, and use those. And also use a
"standard" set of system paths.

But Unifont is a different animal: Gargoyle installs it, and wants to
find its Unifont, and its Unifont only. The user shouldn't be able to
specify where it is.

Now Unifont and themes are searched differently. Themese still use the
concept of "data directories", which on Unix systems means following XDG
(via Qt's QStandardPaths), or some platform-specific hackery for Windows
and Mac DMG files.

But Unifont is installed to a known location, and found directly. For
Unix systems, CMake knows where it installs it, and directly tells
Gargoyle where that is. As usual with Gargoyle on Windows, the location
is the same as the Gargoyle binary. And as usual on Mac, it's in a
specific location in the DMG bundle.