Clean up Windows checks
MINGW used to be used when checking for Windows builds, because it was
the only supported Windows toolchain. Now MSVC (via Clang, at least) is
supported, so a lot of the Windows checks became "MINGW OR MSVC", but at
this point the toolchain isn't important, the OS is; so change such
tests to WIN32.

This also removes a couple of vestigial OS checks which specifically
error out if the OS is unknown. That's unnecessary because CMake will
check for necessary libraries, failing if they can't be found.