Fix, to some degree, issues with scaling
The scalers were added for Infocom games, and really shouldn't be used
elsewhere. Modern games are designed with higher resolutions in mind,
and use tricks like taking a 1x1 pixel and blowing it up to cover an
arbitrary amount of space. This doesn't play well with xBRZ, so avoid
it, at least somewhat. Also update the documentation in garglk.ini to
suggest to the user not to enable scaling for non-Infocom games.