Support the showing of a caret (cursor) in text grid windows
Glk specifically notes that grid windows won't necessarily show a cursor
(see §3.5.4), so Gargoyle's traditional behavior isn't wrong. However,
the license application at the beginning of the game Bureaucracy is
extremely frustrating to use without a visible cursor. While such
frustration actually dovetails nicely with the goal of the game,
Infocom's interpreters weren't so cruel, and did provide a cursor. So do
that here as well. The game My Angel also benefits, as it does line
input in the upper window.

This feature is optional, though enabled by default. The "gridcaret"
config option allows it to be disabled if the traditional behavior is
desired.

This also introduces a small change in how the focused (for input)
window is chosen. Previously Gargoyle would, every time a key was
pressed, search for a window requesting input and, upon finding one,
make it the currently-focused window. Caret-drawing functions then check
whether they are in a focused window before drawing, so that, if there
are multiple inputs being requested, for example, only one window
actually draws the cursor.

This reactive behavior can cause the cursor to _not_ be displayed in
some circumstances. This was immediately obvious with Bureaucracy: the
main buffer window had focus, and when the upper window was selected and
input requested, the cursor was _not_ drawn, because no input had yet
occurred. Once a key was pressed, the cursor showed up.

But this is not a new problem. Current Gargoyle has the same issue. It's
just not really noticeable, generally, since most games don't flip back
and forth between windows while doing input on both.

Consider the following Inform program:

    Array text -> 63;
    Array parse -> 42;

    [Main;
        print "Hit a key...";
        @split_window 1;
        @set_window 1;
        @read_char 1 -> sp;
        @set_window 0;
        text->0 = 60;
        parse->0 = 10;
        print "^> ";
        read text parse;
    ];

In current Gargoyle, once window 0 is selected and `read` is called, the
cursor _is not visible_, because the focus was on window 1 due to the
@read_char call. When the user hit a key at @read_char, window 1 got
focus. Then when window 0 was re-selected, it _did not_ get focus,
because of the previously-noted reactive nature of focus selection.

Now whenever input is requested, the window on which the input is
requested becomes focused. This does change behavior slightly:

    winid_t mainwin = glk_window_open(0, 0, 0, wintype_TextBuffer, 0);
    winid_t upperwin = glk_window_open(mainwin, winmethod_Above | winmethod_Fixed, 10, wintype_TextGrid, 0);

    char mainbuf[30];
    char upperbuf[30];

    glk_request_line_event(mainwin, mainbuf, sizeof mainbuf, 0);

    event_t event;
    do {
        glk_select(&event);
    } while (event.type != evtype_LineInput);

    glk_request_line_event(mainwin, mainbuf, sizeof mainbuf, 0);
    glk_request_line_event(upperwin, upperbuf, sizeof upperbuf, 0);

    while (true) {
        glk_select(&event);
    }

In the original Gargoyle code, `mainwin` retains focus. When a key is
pressed, Gargoyle sees that the currently-focused window (`mainwin`) has
pending input, so it maintains focus. The new Gargoyle code will
forcibly set the focus window to `upperin` upon request for input on it.
At this point, before any keypresses, `upperwin` becomes the focused
window.

What that means is that in the old Gargoyle, after the initial line
input, when there are now two simultaneous line inputs, the user will be
typing in the main window. But in the new Gargoyle, the user will be
typing in the upper window. Glk makes no statements one way or the other
about how this ought to work, leaving it up to the implementation.