Fix style hint clearing
Style clearing works by reverting to the "default" style, but the
default style was fixed to _Gargoyle's_ default, not the user's setting.
If the user changed colors from the default, then clearing would _not_
go back to his preferred settings. Now the defaults are updated to the
user's preferred style settings.