Improve parsing of numbers in the config file
The impetus for this change is C's (and thus C++'s) absolutely bonkers
locale situation. By default, the C locale is used, even if the user
sets $LANG (or related variables). A program can decide to be locale
aware by calling setlocale(). This sets the global locale, which in
isolation sounds OK, since otherwise you'd need to specify a locale
every single time you use a locale-aware function. But there are a
couple problems:

* Any shared libraries that are loaded run in the same space as the
  program, meaning if _they_ call setlocale(), _your_ program is
  affected, since it's global.
* Relatedly, this means that if you're using a locale-aware function,
  and a library has an active thread that calls setlocale(), you've got
  a race and you can't do a single thing about it.

Gargoyle has never called setlocale(). But Qt _does_ call setlocale(),
meaning that on most platforms, Gargoyle becomes locale-aware, without
meaning to.

The most immediately relevant result of this is that if your locale has
a comma as the decimal separator, the default config file will not work
properly, as it uses periods.

To work around this, Gargoyle now parses strings into numbers using
streams (specifically a std::istringstream); that's because streams can
have their locales set independently. Gargoyle parses the strings in the
classic (C) locale, which means that the period is _always_ the decimal
separator now, regardless of the user's locale.

But since earlier releases required a comma in some locales, it's
possible users have commas in their config files. So Gargoyle just
treats commas as periods, regardless of locale.

Gargoyle also used to reject doubles containing [eE], for reasons I no
longer remember. I don't see any reason to reject scientific notation in
the config file, so now it's allowed.

As a side note, Gargoyle also used to accept "inf" and "nan" as
floating-point values: not intentionally, but accidentally, because
std::stod (via std::strtod) accepts them. operator>> also uses
std::strtod(), but it _does not_ support "inf" or "nan", so now Gargoyle
doesn't either. Why doesn't operator>> support these values, despite
using std::strtod()? Because operator>> uses std::get_num(), which
stops lexing on a character that doesn't match this whitelist:

    0123456789abcdefxABCDEFX+-

"inf" and "nan" contain non-whitelisted characters, so never get passed
to std::strtod(). But that list is for C++14. In C++23, the list is now:

    0123456789abcdefpxABCDEFPX+-

OK, so "inf" and "nan" still won't parse, but it's conceivable that at
some point in the future, "n" will get added, at which point "nan" will
become parseable. Will such a standard then specifically call out "nan"
as invalid, to be compatible with older standards? Who knows!

I add all this information because I spent a long time scouring the C++
standard and I will forget it if I don't write it down somewhere, and
this commit message is already so long that nobody will get to the end,
so this seemed a fitting place for it.