Protect some NOMINMAX definitions
Newer versions of gcc define NOMINMAX when some C++-specific headers are
included. This causes warnings when Gargoyle (re)defines NOMINMAX. It
really shouldn't since NOMINMAX is not a reserved identifier and
Gargoyle should be able to define it as much as it wants, i.e., the
implementation should _not_ be defining names that are outside of its
namespace. But blame Windows: it puts "min" and "max" macros in the
global namespace, completely breaking std::min and std::max, and gcc
apparently (and reasonably) decided that unbreaking C++ on Windows is
worth it.

Not all uses of NOMINMAX are protected here: if it doesn't create an
_actual_ problem, I've left it alone, since I'm not even sure I like
this particular approach. Other possibilities:

* Always define NOMINMAX _before_ any headers are included (doesn't save
  us when it's defined _in_ a header itself, though).
* Set NOMINMAX via CMake (too broad, probably, although in reality it
  won't matter).

For now, at least, this silences annoying warnings.