Support click to position cursor and deleting highlighted chunks
This is _slightly_ unintuitive, in a sense: the "input box", so to
speak, isn't really any different from the main text. It's kind of like
a terminal emulator in that way, and terminal emulators tend not to
support highlighting input and cursor positioning via mouse; but this is
also a GUI program, so it probably ought to behave like one.

The end result is something that feels a little off no matter whether
you're more comfortable with terminals or GUIs; but it's probably a move
in the right direction, and will likely be followed up with even more
line-editing niceties.