Improve key sequence handling in the Qt frontend
The impetus for this was allowing Ctrl-Q to quit. This could have just
been added as Qt::ControlModifier + Qt::Key_Q which would have been
pretty good globally, as ControlModifier on Mac actually maps to
Command; but when I learned that, it made it obvious that key handling
was woefully broken, as things like Ctrl-a (Emacs bindind for going to
the beginning of the line) are mapped to Cmd-a, which makes no sense.

Qt has the concept of "key sequences" which are higher level concepts
like copy, paste, quit, etc, which bind to platform-specific key
sequences. That's great, as it allows you to use QKeySequence::Quit and
the platform can define what that means! Except... Qt doesn't define a
Quit action for Windows...

Anyway, this is an attempt at being platform agnostic while still being
predictable.

The whole broken Mac thing doesn't actually matter much because as of
right now, Mac uses its own front-end code. But when Qt eventually
replaces the Mac front-end, it'd be nice if it actually worked right.